+++ Common/src/main/java/at/petrak/hexcasting/client/RegisterClientStuff.java
@@ -114,6 +114,7 @@ public class RegisterClientStuff {
         registerWandOverrides(HexItems.STAFF_EDIFIED.get());
         // purposely skip quenched
         registerWandOverrides(HexItems.STAFF_MINDSPLICE.get());
+        registerWandOverrides(HexItems.STAFF_NEURAL.get());
 
         registerGaslight4(HexItems.STAFF_QUENCHED.get());
         registerGaslight4(HexBlocks.QUENCHED_ALLAY.get().asItem());
@@ -128,6 +129,7 @@ public class RegisterClientStuff {
         x.setRenderLayer(HexBlocks.EDIFIED_TRAPDOOR.get(), RenderType.cutout());
         x.setRenderLayer(HexBlocks.AKASHIC_BOOKSHELF.get(), RenderType.cutout());
         x.setRenderLayer(HexBlocks.SCONCE.get(), RenderType.cutout());
+        x.setRenderLayer(HexBlocks.NEURAL_MESH.get(), RenderType.cutout());
 
         x.setRenderLayer(HexBlocks.AMETHYST_EDIFIED_LEAVES.get(), RenderType.cutoutMipped());
         x.setRenderLayer(HexBlocks.AVENTURINE_EDIFIED_LEAVES.get(), RenderType.cutoutMipped());