+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 - Updated to Minecraft 1.21.1 ([#985](https://github.com/FallingColors/HexMod/pull/985)) @SuperKnux @slava110
 - Added Neural Fiber, a new endgame material harvested from villagers using a modified mindflaying process ([#1296](https://github.com/FallingColors/HexMod/pull/1296)) @Robotgiggle @Falkory220
+- Added Neural Mesh, a multipart block which can be attached to Budding Amethyst to make it grow faster ([#1297](https://github.com/FallingColors/HexMod/pull/1297)) @Robotgiggle
 - Added Grand Caster Robes, a hex-based armor set with a variety of abilities ([#1296](https://github.com/FallingColors/HexMod/pull/1296)) @Robotgiggle @Falkory220
   - Provides defense values midway between iron and diamond
   - Provides the scrying lens effect, increased ambit, and a casting cost discount